requirements.txt
anthropic==0.52.0
matplotlib==3.10.3
numpy==2.2.6
openai==1.82.0
pandas==2.2.3
PyYAML==6.0.2
seaborn==0.13.2
